% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrosoftAdsR.R
\name{Authentication}
\alias{Authentication}
\title{Microsoft Ads Authentication Function}
\usage{
Authentication(credentials)
}
\arguments{
\item{Credentials}{Pass the client id, account id, customer id and dev token}
}
\description{
This function allows you to authenticate with Microsoft Ads
}
\examples{
Authentication()
}
\keyword{Authentication}
