% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrosoftAdsR.R
\name{getDownloadUrl}
\alias{getDownloadUrl}
\title{Download Report Function}
\usage{
getDownloadUrl(credentials, reportId)
}
\arguments{
\item{Credentials}{Pass the access token, developer token and account id along with the report id}
}
\description{
This function uses the report ID - polls to check if the report is ready then hits the download url
}
\examples{
getDownloadUrl()
}
\keyword{getDownloadUrl}
