% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrosoftAdsR.R
\name{getReportId}
\alias{getReportId}
\title{Get Report ID Function}
\usage{
getReportId(credentials, report, columns, startDate, endDate)
}
\arguments{
\item{Credentials}{Pass the client id, account id, customer id and dev token}
}
\description{
This function allows you to request the report and get the required report id
}
\examples{
getReportId()
}
\keyword{getReportId}
